backend/requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# OpenAI
openai>=1.3.0

# Data
akshare>=1.12.0
pandas>=2.0.0
numpy>=1.24.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0

# Template
jinja2>=3.1.0

# Optional: Technical Analysis
# ta-lib>=0.4.28  # 需额外安装